move utility functions out of index.js
