+++ lib/rules/named-grid-areas-no-invalid/index.js
@@ -4,6 +4,8 @@
 
 const _ = require('lodash');
 const declarationValueIndex = require('../../utils/declarationValueIndex');
+const findNotContiguousOrRectangular = require('./utils/findNotContiguousOrRectangular');
+const isRectangular = require('./utils/isRectangular');
 const report = require('../../utils/report');
 const ruleMessages = require('../../utils/ruleMessages');
 const validateOptions = require('../../utils/validateOptions');
@@ -71,56 +73,6 @@ function rule(actual) {
 			}
 		});
 	};
-}
-
-function columnsPerRow(areas) {
-	return areas.map((row) => row.length);
-}
-
-function isRectangular(areas) {
-	return columnsPerRow(areas).every((val, i, arr) => val === arr[0]);
-}
-
-function isContiguousAndRectangular(areas, name) {
-	const indicesByRow = areas.map((row) => {
-		const indices = [];
-		let idx = row.indexOf(name);
-
-		while (idx !== -1) {
-			indices.push(idx);
-			idx = row.indexOf(name, idx + 1);
-		}
-
-		return indices;
-	});
-
-	for (let i = 0; i < indicesByRow.length; i++) {
-		for (let j = i + 1; j < indicesByRow.length; j++) {
-			if (indicesByRow[i].length === 0 || indicesByRow[j].length === 0) {
-				continue;
-			}
-
-			if (!_.isEqual(indicesByRow[i], indicesByRow[j])) {
-				return false;
-			}
-		}
-	}
-
-	return true;
-}
-
-function namedAreas(areas) {
-	const names = new Set(_.flatten(areas));
-
-	names.delete('.');
-
-	return Array.from(names);
-}
-
-function findNotContiguousOrRectangular(areas) {
-	return Array.from(
-		new Set(namedAreas(areas).filter((name) => !isContiguousAndRectangular(areas, name))),
-	);
 }
 
 rule.ruleName = ruleName;
